Update and rename welcome.py# to welcome.py

+++ SONALI/plugins/tools/welcome.py
@@ -1,25 +1,10 @@
 from SONALI import app
-from pyrogram import Client, filters
-from pyrogram.enums import ChatMemberStatus
-from pyrogram.errors import (
-    ChatAdminRequired,
-    InviteRequestSent,
-    UserAlreadyParticipant,
-    UserNotParticipant,
-)
 from pyrogram.errors import RPCError
 from pyrogram.types import ChatMemberUpdated, InlineKeyboardMarkup, InlineKeyboardButton
 from os import environ
 from typing import Union, Optional
 from PIL import Image, ImageDraw, ImageFont
 from os import environ
-import requests
-import random
-from SONALI import app, userbot
-from SONALI.misc import SUDOERS
-from pyrogram import *
-from pyrogram.types import *
-from SONALI.utils.RAUSHAN_ban import admin_filter
 import random
 from pyrogram import Client, filters
 from pyrogram.types import ChatJoinRequest, InlineKeyboardButton, InlineKeyboardMarkup
@@ -30,12 +15,8 @@ from PIL import Image, ImageDraw, ImageFont, ImageEnhance
 from asyncio import sleep
 from pyrogram import filters, Client, enums
 from pyrogram.enums import ParseMode
-from pyrogram import *
-from pyrogram.types import *
 from logging import getLogger
-from SONALI.utils.istkhar_ban import admin_filter
-import os
-from SONALI.misc import SUDOERS
+from SONALI.utils.RAUSHAN_ban import admin_filter
 from PIL import ImageDraw, Image, ImageFont, ImageChops
 from pyrogram import *
 from pyrogram.types import *
@@ -52,8 +33,10 @@ random_photo = [
 # --------------------------------------------------------------------------------- #
 
 
-LOGGER = getLogger(__name__)
 
+
+
+LOGGER = getLogger(__name__)
 
 class WelDatabase:
     def __init__(self):
@@ -70,9 +53,7 @@ class WelDatabase:
         if chat_id in self.data:
             del self.data[chat_id]
 
-
 wlcm = WelDatabase()
-
 
 class temp:
     ME = None
@@ -81,6 +62,7 @@ class temp:
     MELCOW = {}
     U_NAME = None
     B_NAME = None
+
 
 
 def circle(pfp, size=(500, 500), brightness_factor=10):
@@ -95,86 +77,21 @@ def circle(pfp, size=(500, 500), brightness_factor=10):
     pfp.putalpha(mask)
     return pfp
 
-
 def welcomepic(pic, user, chatname, id, uname, brightness_factor=1.3):
-    background = Image.open("ISTKHARMUSIC/assets/wel2.png")
+    background = Image.open("SONALI/assets/wel2.png")
     pfp = Image.open(pic).convert("RGBA")
-    pfp = circle(pfp, brightness_factor=brightness_factor)
-    pfp = pfp.resize((892, 880))
+    pfp = circle(pfp, brightness_factor=brightness_factor) 
+    pfp = pfp.resize((500, 500))
     draw = ImageDraw.Draw(background)
-    font = ImageFont.truetype("ISTKHARMUSIC/assets/font.ttf", size=95)
-    welcome_font = ImageFont.truetype("ISTKHARMUSIC/assets/font.ttf", size=45)
+    font = ImageFont.truetype('SONALI/assets/font.ttf', size=60)
+    welcome_font = ImageFont.truetype('SONALI/assets/font.ttf', size=60)
+    
+ #   draw.text((630, 230), f"USERNAME : {uname}", fill=(255, 255, 255), font=font)
+   # draw.text((630, 300), f'NAME: {user}', fill=(255, 255, 255), font=font)
+    draw.text((630, 450), f'ID: {id}', fill=(255, 255, 255), font=font)
 
-    # Draw user's name with shining red fill and dark saffron border
-    draw.text((1770, 1015), f": {user}", fill=(255, 0, 0), font=font)
-    draw.text(
-        (1770, 1015),
-        f": {user}",
-        fill=None,
-        font=font,
-        stroke_fill=(255, 153, 51),
-        stroke_width=6,
-    )
-
-    # Draw user's id with shining blue fill and white border
-    draw.text((1530, 1230), f": {id}", fill=(0, 0, 139))
-    draw.text(
-        (1530, 1230),
-        f": {id}",
-        fill=None,
-        font=font,
-        stroke_fill=(255, 255, 255),
-        stroke_width=0,
-    )
-
-    # Draw user's username with white fill and green border
-    draw.text((2030, 1450), f": {uname}", fill=(255, 255, 255), font=font)
-    draw.text(
-        (2030, 1450),
-        f": {uname}",
-        fill=None,
-        font=font,
-        stroke_fill=(0, 128, 0),
-        stroke_width=6,
-    )
-
-    # Resize photo and position
-    pfp_position = (255, 323)
+    pfp_position = (48, 88)
     background.paste(pfp, pfp_position, pfp)
-
-    # Calculate circular outline coordinates
-    center_x = pfp_position[0] + pfp.width / 2
-    center_y = pfp_position[1] + pfp.height / 2
-    radius = min(pfp.width, pfp.height) / 2
-
-    # Draw circular outlines
-    draw.ellipse(
-        [
-            (center_x - radius - 10, center_y - radius - 10),
-            (center_x + radius + 10, center_y + radius + 10),
-        ],
-        outline=(255, 153, 51),
-        width=25,
-    )  # Saffron border
-
-    draw.ellipse(
-        [
-            (center_x - radius - 20, center_y - radius - 20),
-            (center_x + radius + 20, center_y + radius + 20),
-        ],
-        outline=(255, 255, 255),
-        width=25,
-    )  # White border
-
-    draw.ellipse(
-        [
-            (center_x - radius - 30, center_y - radius - 30),
-            (center_x + radius + 30, center_y + radius + 30),
-        ],
-        outline=(0, 128, 0),
-        width=25,
-    )  # Green border
-
     background.save(f"downloads/welcome#{id}.png")
     return f"downloads/welcome#{id}.png"
 
@@ -197,21 +114,18 @@ async def auto_state(_, message):
                 await message.reply_text("**ᴡᴇʟᴄᴏᴍᴇ ɴᴏᴛɪғɪᴄᴀᴛɪᴏɴ ᴀʟʀᴇᴀᴅʏ ᴅɪsᴀʙʟᴇᴅ !**")
             else:
                 await wlcm.add_wlcm(chat_id)
-                await message.reply_text(
-                    f"**ᴅɪsᴀʙʟᴇᴅ ᴡᴇʟᴄᴏᴍᴇ ɴᴏᴛɪғɪᴄᴀᴛɪᴏɴ ɪɴ** {message.chat.title}"
-                )
+                await message.reply_text(f"**ᴅɪsᴀʙʟᴇᴅ ᴡᴇʟᴄᴏᴍᴇ ɴᴏᴛɪғɪᴄᴀᴛɪᴏɴ ɪɴ** {message.chat.title}")
         elif state == "on":
             if not A:
                 await message.reply_text("**ᴇɴᴀʙʟᴇ ᴡᴇʟᴄᴏᴍᴇ ɴᴏᴛɪғɪᴄᴀᴛɪᴏɴ.**")
             else:
                 await wlcm.rm_wlcm(chat_id)
-                await message.reply_text(
-                    f"**ᴇɴᴀʙʟᴇᴅ ᴡᴇʟᴄᴏᴍᴇ ɴᴏᴛɪғɪᴄᴀᴛɪᴏɴ ɪɴ ** {message.chat.title}"
-                )
+                await message.reply_text(f"**ᴇɴᴀʙʟᴇᴅ ᴡᴇʟᴄᴏᴍᴇ ɴᴏᴛɪғɪᴄᴀᴛɪᴏɴ ɪɴ ** {message.chat.title}")
         else:
             await message.reply_text(usage)
     else:
         await message.reply("**sᴏʀʀʏ ᴏɴʟʏ ᴀᴅᴍɪɴs ᴄᴀɴ ᴇɴᴀʙʟᴇ ᴡᴇʟᴄᴏᴍᴇ ɴᴏᴛɪғɪᴄᴀᴛɪᴏɴ!**")
+
 
 
 @app.on_chat_member_updated(filters.group, group=-3)
@@ -223,16 +137,16 @@ async def greet_new_member(_, member: ChatMemberUpdated):
         return
 
     user = member.new_chat_member.user if member.new_chat_member else member.from_user
-
+    
     # Add the modified condition here
-    if member.new_chat_member and not member.old_chat_member:
-
+    if member.new_chat_member and not member.old_chat_member and member.new_chat_member.status != "kicked":
+    
         try:
             pic = await app.download_media(
                 user.photo.big_file_id, file_name=f"pp{user.id}.png"
             )
         except AttributeError:
-            pic = "ISTKHARMUSIC/assets/upic.png"
+            pic = "SONALI/assets/upic.png"
         if (temp.MELCOW).get(f"welcome-{member.chat.id}") is not None:
             try:
                 await temp.MELCOW[f"welcome-{member.chat.id}"].delete()
@@ -243,31 +157,30 @@ async def greet_new_member(_, member: ChatMemberUpdated):
                 pic, user.first_name, member.chat.title, user.id, user.username
             )
             button_text = "๏ ᴠɪᴇᴡ ɴᴇᴡ ᴍᴇᴍʙᴇʀ ๏"
-            add_button_text = "๏ ᴋɪᴅɴᴀᴘ ᴍᴇ ๏"
+            add_button_text = "✙ ᴋɪᴅɴᴀᴘ ᴍᴇ ✙"
             deep_link = f"tg://openmessage?user_id={user.id}"
             add_link = f"https://t.me/{app.username}?startgroup=true"
             temp.MELCOW[f"welcome-{member.chat.id}"] = await app.send_photo(
                 member.chat.id,
                 photo=welcomeimg,
                 caption=f"""
-**💠────✦ ᴡᴇʟᴄᴏᴍᴇ ✦────💠**
+**⎊─────☵ ᴡᴇʟᴄᴏᴍᴇ ☵─────⎊**
 
-▰▰▰▰▰▰▰▰▰▰▰▰▰
-**▫️ ɴᴀᴍᴇ »** {user.mention}
-**🔸 ɪᴅ »** `{user.id}`
-**▪️ ᴜ_ɴᴀᴍᴇ »** @{user.username}
-**▫️ ᴛᴏᴛᴀʟ ᴍᴇᴍʙᴇʀs »** {count}
-**▫️ ᴍᴀᴅᴇ ʙʏ »** [ꪜ 𝛊 ɭ ɭ ᧘ 𝛊 𝛈](https://t.me/iamakki001)
-▰▰▰▰▰▰▰▰▰▰▰▰▰
+**▬▭▬▭▬▭▬▭▬▭▬▭▬▭▬**
 
-**💠─────✧❅✦❅✧─────💠**
+**☉ ɴᴀᴍᴇ ⧽** {user.mention}
+**☉ ɪᴅ ⧽** `{user.id}`
+**☉ ᴜ_ɴᴀᴍᴇ ⧽** @{user.username}
+**☉ ᴛᴏᴛᴀʟ ᴍᴇᴍʙᴇʀs ⧽** {count}
+
+**▬▭▬▭▬▭▬▭▬▭▬▭▬▭▬**
+
+**⎉──────▢✭ 侖 ✭▢──────⎉**
 """,
-                reply_markup=InlineKeyboardMarkup(
-                    [
-                        [InlineKeyboardButton(button_text, url=deep_link)],
-                        [InlineKeyboardButton(text=add_button_text, url=add_link)],
-                    ]
-                ),
+                reply_markup=InlineKeyboardMarkup([
+                    [InlineKeyboardButton(button_text, url=deep_link)],
+                    [InlineKeyboardButton(text=add_button_text, url=add_link)],
+                ])
             )
         except Exception as e:
-            return
+            LOGGER.error(e)
